separating logic of task to irrigate, from valves control
